[feature][implementation] add room_auth table and unit test
