+++ .gitignore
@@ -2,7 +2,7 @@ backend/src/base/__pycache__/
 backend/src/library/__pycache__/
 backend/src/platform/douyin/__pycache__/
 backend/src/database/__pycache__/
-backend/src/database/database_table/__pycache__
+backend/src/database/table/__pycache__
 backend/src/platform/__pycache__/
 backend/src/platform/other/__pycache__
 venv/
