+++ backend/src/database/table/favorite_owner.py
