+++ backend/src/database/table/social_media_stream_db_table.py
